move PatchUrlResolver out from Formula dir
