+++ lib/PatchUrlResolver.rb
